+++ paper_uploads/cloudinary/models/file.py
@@ -24,7 +24,7 @@ class CloudinaryFile(BacklinkModelMixin, CloudinaryFileFieldResource):
 
     def save(self, *args, **kwargs):
         if not self.pk and not self.display_name:
-            self.display_name = self.basename
+            self.display_name = self.resource_name
         super().save(*args, **kwargs)
 
     def get_file(self) -> Optional[CloudinaryFieldFile]:
@@ -42,7 +42,7 @@ class CloudinaryFile(BacklinkModelMixin, CloudinaryFileFieldResource):
     def as_dict(self) -> Dict[str, Any]:
         return {
             **super().as_dict(),
-            "name": self.display_name or self.basename,
+            "name": self.display_name or self.resource_name,
             "file_info": "({ext}, {size})".format(
                 ext=self.extension, size=filesizeformat(self.size)
             ),